pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avro"
version = "0.1.0"
description = "Aerodrome veAERO vote optimizer — V1 recommender CLI."
requires-python = ">=3.10"
dependencies = [
    "sugar-sdk==0.3.1",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
avro = "avro.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"
